Refactor some components and implement tracks release in the current year you've listened to.
